OLD_TRY/requirements.txt
# Core scientific computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web requests for ESMFold API
requests>=2.31.0

# YAML parsing (for Boltz input)
pyyaml>=6.0

# Progress bars
tqdm>=4.65.0

# BioPython for sequence handling
biopython>=1.81

# Numerical optimization
scikit-learn>=1.3.0

# Additional utilities
joblib>=1.3.0
networkx>=3.1

# Protein language models and embeddings
transformers>=4.30.0

# Structure file handling
prody>=2.4.0

# NOTE: The following packages require special installation steps:
#
# 1. PyTorch (MUST be installed first, separately):
#    pip install torch torchvision torchaudio
#    Or for GPU: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
#
# 2. Chemistry packages (recommend using conda):
#    conda install -c conda-forge rdkit openmm pdbfixer
#
# 3. PyTorch Geometric packages (install AFTER PyTorch):
#    pip install torch-geometric torch-scatter torch-sparse torch-cluster
#
# 4. ESM (install AFTER PyTorch):
#    pip install fair-esm
#
# 5. Docking tools (may need manual installation):
#    - Boltz: pip install git+https://github.com/jwohlwend/boltz.git
#    - DiffDock: git clone and pip install -e
#    - Chai-1: pip install chai-lab
#    - ProtoBind: Check official repository
